requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
certifi==2025.1.31
crispy-bootstrap5==2025.4
Django==5.2
django-allauth==65.7.0
django-apscheduler==0.7.0
django-auditlog==3.0.0
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-environ==0.12.0
django-filter==25.1
django-haystack==3.3.0
django-mathfilters==1.0.0
django-simple-history==3.8.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
elastic-transport==8.17.1
elasticsearch==8.17.2
packaging==24.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
Whoosh==2.7.4
